refactor(threads): split the transcript migration out of thread_index
f-29: `thread_index.rs` had grown to 1,202 lines by carrying a second,
unrelated concern — a transactional transcript-rebuild state machine with
its own retry outcome, conflict classifier, backoff loop, page transaction
and completion marker — alongside the thread-listing projection it owns.
Moved to `transcript_migration.rs` (thread_index 1202 -> 941 lines); no
behavior change.
